+++ packages/orq-rc/src/orq_ai_sdk/models/retrievedatasetop.py
@@ -81,7 +81,7 @@ class RetrieveDatasetResponseBody(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2026-03-30T07:57:15.372Z")
+    updated: Optional[datetime] = parse_datetime("2026-03-30T08:19:08.752Z")
     r"""The date and time the resource was last updated"""
 
     @model_serializer(mode="wrap")