+++ packages/orq-rc/RELEASES.md
@@ -6679,3 +6679,13 @@ Based on:
 - [python v4.7.0-rc.27] packages/orq-rc
 ### Releases
 - [PyPI v4.7.0-rc.27] https://pypi.org/project/orq-ai-sdk/4.7.0-rc.27 - packages/orq-rc
+
+## 2026-03-30 08:22:30
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.759.3 (2.869.25) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v4.7.0-rc.28] packages/orq-rc
+### Releases
+- [PyPI v4.7.0-rc.28] https://pypi.org/project/orq-ai-sdk/4.7.0-rc.28 - packages/orq-rc